refactor(api): split middleware/auth.ts into roles/context/guards/jwt modules (#724)
middleware/auth.ts was 419 lines (over the 400 soft cap), co-locating five
separable concerns. Extract them into cohesive auth/ modules, leaving
middleware/auth.ts as the thin requireAuth middleware + context accessors plus a
barrel that re-exports the prior public surface so all ~190 import sites compile
unchanged.

- auth/roles.ts:   AuthUser, STAFF/FLEET_WRITE/MANAGEMENT_READ aliases,
                   isOperatorRole / isValidRole / isAuthUser
- auth/context.ts: CallerContext + SYSTEM/PUBLIC contexts + toCallerContext
- auth/guards.ts:  Forbidden/OperatorRequiredError + the six require* guards
- auth/jwt.ts:     session-token constants/types + verify/mint/verifyApiKey crypto

Behavior-preserving: api typecheck clean across all import sites; 1297 unit tests
green; lint:size no longer warns for these files; lint:boundaries OK.
