+++ packages/api/src/auth/guards.ts
@@ -1,0 +1,110 @@
+import { OPERATOR_ROLES, PLATFORM_ROLES } from '@kuruma/shared/auth/roles'
+
+import type { CallerContext } from './context'
+import { FLEET_WRITE_ROLES, MANAGEMENT_READ_ROLES } from './roles'
+
+/**
+ * Thrown by repo-layer guards when a non-authorised caller hits a
+ * protected method. The global error handler maps this to a 403 response
+ * so a bypassed route-level gate surfaces as a policy denial, not a 500.
+ */
+export class ForbiddenError extends Error {
+  readonly name = 'ForbiddenError'
+  constructor(message = 'Forbidden') {
+    super(message)
+  }
+}
+
+/**
+ * Thrown when a non-operator caller (PLATFORM_ADMIN / legacy STAFF / ADMIN) tries
+ * to create tenant-owned inventory without naming a target operator, and the
+ * target cannot be inferred (zero or 2+ operators exist). Replaces the old
+ * silent Best-Car-Rental default (#401) so a legacy admin write can no longer be
+ * misattributed once a second operator exists. Mapped to 422 by the global
+ * handler — the request is well-formed but missing a required `operatorId`.
+ */
+export class OperatorRequiredError extends Error {
+  readonly name = 'OperatorRequiredError'
+  constructor(message = 'operatorId is required') {
+    super(message)
+  }
+}
+
+/**
+ * Repo-layer read guard for operator-private config. Must be called BEFORE
+ * `operatorReadScope(ctx)`, because that helper maps every non-operator role —
+ * including RENTER — to `{kind:'all'}` (the vehicle catalog is public). Without
+ * this seal a renter or PARTNER could read every operator's insurance/fees
+ * config. Throws `ForbiddenError` (-> 403) for RENTER / PARTNER (slice-4 [P0]).
+ */
+export function requireManagementRead(ctx: CallerContext): void {
+  if (!MANAGEMENT_READ_ROLES.has(ctx.role)) {
+    throw new ForbiddenError('management read scope required')
+  }
+}
+
+/**
+ * Gate for PLATFORM-level reads that span every operator — the #462 admin
+ * revenue tab. Admits only the platform tier (`PLATFORM_ROLES` = {PLATFORM_ADMIN}
+ * after #487), the exact set the web `_admin` portal admits. References
+ * PLATFORM_ROLES directly (not the STAFF_ROLES alias) so #487's tightening of
+ * PLATFORM_ROLES → {PLATFORM_ADMIN} narrows this gate automatically. Deliberately
+ * EXCLUDES OPERATOR_* (a tenant must never see another partner's revenue) and
+ * RENTER / PARTNER. Narrower than `bypassScope`, which also covers PARTNER
+ * (Trip.com) — a 3rd-party caller must not read revenue.
+ */
+export function requirePlatformRead(ctx: CallerContext): void {
+  if (!PLATFORM_ROLES.has(ctx.role)) {
+    throw new ForbiddenError('platform admin scope required')
+  }
+}
+
+/**
+ * Repo-layer guard for fleet mutation methods. Admits STAFF roles and
+ * tenant-scoped operators (`FLEET_WRITE_ROLES`); a tenant-scoped caller missing
+ * its operatorId fails closed via `requireOperatorScope`. Defence in depth
+ * against a route forgetting its gate (issue #329). The caller's tenant is
+ * enforced by the repository's operator predicate, so an admitted operator can
+ * only mutate its own vehicles. `SYSTEM_CONTEXT` (PLATFORM_ADMIN) passes.
+ */
+export function requireFleetWriteScope(ctx: CallerContext): void {
+  if (!FLEET_WRITE_ROLES.has(ctx.role)) {
+    throw new ForbiddenError('fleet write scope required')
+  }
+  requireOperatorScope(ctx)
+}
+
+/**
+ * Repo-layer guard for operator-scoped paths. Throws `ForbiddenError` if a
+ * tenant-scoped caller (OPERATOR_*) reached a scoped method without an
+ * operatorId — a fail-closed defence against a token that lost its tenant claim.
+ */
+export function requireOperatorScope(ctx: CallerContext): void {
+  if (OPERATOR_ROLES.has(ctx.role) && !ctx.operatorId) {
+    throw new ForbiddenError('operator scope required')
+  }
+}
+
+/**
+ * Fail-closed guard for repos NOT yet operator-scoped in this slice. An
+ * OPERATOR_* caller hitting such a repo is rejected with a specific message
+ * rather than silently falling through to a global-bypass path (plan v2 P1a).
+ * Legacy STAFF/ADMIN, PARTNER, PLATFORM_ADMIN, and RENTER paths are unaffected.
+ */
+export function rejectOperatorContextUntilScoped(ctx: CallerContext, repoName: string): void {
+  if (OPERATOR_ROLES.has(ctx.role)) {
+    throw new ForbiddenError(`${repoName} not yet operator-scoped`)
+  }
+}
+
+/**
+ * Guard for platform-admin-only paths (operator bootstrap, proposal §9 item 23).
+ * Only PLATFORM_ADMIN passes — legacy STAFF/ADMIN do NOT, since operator
+ * creation is a platform-governance action, not a fleet-management one.
+ * `SYSTEM_CONTEXT` (role PLATFORM_ADMIN) passes.
+ */
+export function requirePlatformAdmin(ctx: CallerContext): void {
+  if (ctx.role !== 'PLATFORM_ADMIN') {
+    throw new ForbiddenError('PLATFORM_ADMIN role required')
+  }
+}
